+++ source/extensions/isaacsim.sensors.physics.ui/isaacsim/sensors/physics/ui/extension.py
@@ -1,0 +1,46 @@
+# Copyright (c) 2022-2024, NVIDIA CORPORATION. All rights reserved.
+#
+# NVIDIA CORPORATION and its licensors retain all intellectual property
+# and proprietary rights in and to this software, related documentation
+# and any modifications thereto. Any use, reproduction, disclosure or
+# distribution of this software and related documentation without an express
+# license agreement from NVIDIA CORPORATION is strictly prohibited.
+#
+import gc
+import weakref
+from functools import partial
+from pathlib import Path
+
+import omni.ext
+import omni.kit.commands
+from isaacsim.core.utils.prims import set_prim_visibility
+from isaacsim.gui.components.menu import make_menu_item_description
+from omni.kit.menu.utils import MenuItemDescription, add_menu_items, remove_menu_items
+from pxr import Gf, Tf
+
+
+class Extension(omni.ext.IExt):
+    def on_startup(self, ext_id: str) -> None:
+
+        menu_items = [
+            make_menu_item_description(ext_id, "Contact Sensor", lambda a=weakref.proxy(self): a._add_contact_sensor()),
+            make_menu_item_description(ext_id, "Imu Sensor", lambda a=weakref.proxy(self): a._add_imu_sensor()),
+        ]
+
+        icon_dir = omni.kit.app.get_app().get_extension_manager().get_extension_path_by_module(__name__)
+        sensor_icon_path = str(Path(icon_dir).joinpath("data/sensor.svg"))
+        self._menu_items = [MenuItemDescription(name="Sensors", glyph=sensor_icon_path, sub_menu=menu_items)]
+        add_menu_items(self._menu_items, "Create")
+
+    def on_shutdown(self):
+        remove_menu_items(self._menu_items, "Create")
+        gc.collect()
+
+    def _get_stage_and_path(self):
+        selectedPrims = omni.usd.get_context().get_selection().get_selected_prim_paths()
+
+        if len(selectedPrims) > 0:
+            curr_prim = selectedPrims[-1]
+        else:
+            curr_prim = None
+        return curr_prim